LayoutTests/inspector/view-events-expected.txt

This tests that events are properly propagated through View hierarchy.

parentView.show(root)
> ParentView.wasShown()
> ParentView.onResize()
parentView.doResize()
> ParentView.onResize()
childView.show()
> ChildView.wasShown()
> ChildView.onResize()
parentView.doResize()
> ParentView.onResize()
> ChildView.onResize()
parentView.hide(root)
> ParentView.willHide()
> ChildView.willHide()
parentView.show(root)
> ParentView.wasShown()
> ParentView.onResize()
> ChildView.wasShown()
> ChildView.onResize()
childView.hide()
> ChildView.willHide()
parentView.hide(root)
> ParentView.willHide()
